Sync contract from address collection
