+++ api/echo.go
@@ -227,6 +227,7 @@ func bind(gr *echo.Group, srv EchoServer) {
 	bindContractAPIs(gr, srv)
 	bindEventAPIs(gr, srv)
 	bindStakingAPIs(gr, srv)
+	bindPrivateAPIs(gr, srv)
 	for _, api := range apis {
 		gr.Add(api.method, api.path, api.fn, api.middlewares...)
 	}
@@ -238,53 +239,6 @@ func bindEventAPIs(gr *echo.Group, srv EchoServer) {
 			method:      echo.DELETE,
 			path:        "/event/duplicate",
 			fn:          srv.RemoveDuplicateEvents,
-			middlewares: nil,
-		},
-	}
-	for _, api := range apis {
-		gr.Add(api.method, api.path, api.fn, api.middlewares...)
-	}
-}
-
-func bindContractAPIs(gr *echo.Group, srv EchoServer) {
-	apis := []restDefinition{
-		{
-			method:      echo.POST,
-			path:        "/contracts",
-			fn:          srv.InsertContract,
-			middlewares: nil,
-		},
-		{
-			method:      echo.PUT,
-			path:        "/contracts",
-			fn:          srv.UpdateContract,
-			middlewares: nil,
-		},
-		{
-			method: echo.GET,
-			// Query params
-			// [?status=(Verified, Unverified)]
-			path:        "/contracts",
-			fn:          srv.Contracts,
-			middlewares: nil,
-		},
-		{
-			method:      echo.GET,
-			path:        "/contracts/:contractAddress",
-			fn:          srv.Contract,
-			middlewares: nil,
-		},
-		{
-			method:      echo.PUT,
-			path:        "/contracts/abi",
-			fn:          srv.UpdateSMCABIByType,
-			middlewares: nil,
-		},
-		{
-			method: echo.GET,
-			// Query params: ?page=0&limit=10&contractAddress=0x&methodName=0x&txHash=0x
-			path:        "/contracts/events",
-			fn:          srv.ContractEvents,
 			middlewares: nil,
 		},
 	}
