+++ db/mgo_test.go
@@ -7,10 +7,54 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"go.uber.org/zap"
 
 	"github.com/kardiachain/kardia-explorer-backend/types"
 )
 
+func SetupTestMGO() (*mongoDB, error) {
+	lgr, err := zap.NewDevelopment()
+	if err != nil {
+		return nil, err
+	}
+	mgoCfg := Config{
+		DbAdapter: "mgo",
+		DbName:    "explorer",
+		URL:       "mongodb://10.10.0.252:27018",
+		MinConn:   1,
+		MaxConn:   4,
+		FlushDB:   false,
+		Logger:    lgr,
+	}
+
+	mgo, err := newMongoDB(mgoCfg)
+	if err != nil {
+		return nil, err
+	}
+	return mgo, nil
+}
+
+func GetMgo() (*mongoDB, error) {
+	lgr, err := zap.NewDevelopment()
+	if err != nil {
+		return nil, err
+	}
+	mgoCfg := Config{
+		DbAdapter: "mgo",
+		DbName:    "explorer",
+		URL:       "mongodb://10.10.0.252:27018",
+		MinConn:   1,
+		MaxConn:   4,
+		FlushDB:   false,
+		Logger:    lgr,
+	}
+
+	mgo, err := newMongoDB(mgoCfg)
+	if err != nil {
+		return nil, err
+	}
+	return mgo, nil
+}
 func Test_mongoDB_GetAddressInfo(t *testing.T) {
 	ctx := context.Background()
 	mgo, err := GetMgo()
